Ensure early exceptions appear in the terminal
Make the process of resetting the terminal and outputting the text of an
unexpected exception to stderr more robust.

It seems resetting the terminal after using dialog races with printing
the text of the exception, and overwrites it with the dialog background.

Sleep for a tiny amount of time to ensure the exception is visible.

This is an ugly workaround and should be replaced.