Re-write tmpdir selection, disable it temporarily
Re-write the procedure of selecting a directory to hold temporary files,
so it does not use an arbitrary list of directories including /mnt
and the user's home directory (which is presumably /root), but starts
with a list of all mountpoints and /tmp, /var/tmp as candidates.
Also fix bug where the creator would choose a directory under a
read-only mount for storing temporary files, and fail.

Still, disable this code path temporarily, and only choose between
/tmp and /var/tmp, if the user has made no explicit choice.
It is very un-intuitive to create temporary files in unrelated
directories without the user's explicit consent.

Perhaps this code path can be enabled again when snf-image-creator
has the ability to propose these alternate locations to the user,
and have them choose explicitly.