Use Python logging, separate logging from Output
Separate the process of logging from the process of interacting with
the user ("Output" instances) to ensure emitting debugging information
is possible from anywhere inside the code.

Use the standard Python logging module for logging, instead of a custom
approach. This brings lots of advantages, including proper log
formatting and timestamping.

Use logging.handlers.SysLogHandler for logging to syslog,
instead of calling the syslog module directly.

Also use a standard default logfile /var/log/snf-image-creator
for persistence, instead of a temporary logfile, and remove redundant
functions show_log(), copy_file(). This simplifies the error handling
path significantly.

The general idea of this commit is that the instances of "Output" should
be independent frontends for interacting with the user, and can be
extended to support Input as well as Output in the future.

This is a Work in Progress. The previous code assumed that all messages
to the user via Output instances would also make it to the log file.
For now, only what is explicitly logged reaches the logfile.

A workaround is to have the base Output class log all messages to the
user, and will be implemented in a future commit.